backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
anthropic>=0.18.0
resend>=0.7.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
markdown>=3.5.0
